The time_cost is the number of iterations of the algorithm, which determines how much time is needed to compute the hash. Increasing this value makes the algorithm slower and more secure against brute-force attacks, but also increases the time required to compute a hash.
The memory_cost is the amount of memory used by the algorithm, which determines how much memory is required to compute the hash. Increasing this value makes the algorithm more secure against memory-based attacks, but also increases the memory required to compute a hash.

parallelism refers to the number of threads used by the algorithm, which determines how many parallel threads are used to compute the hash. Increasing this value makes the algorithm faster, but also increases the risk of side-channel attacks.

the hash_len is the length of the output hash in bytes. Increasing this value makes the output hash longer and more secure against collisions.

Whereas, the salt_len is the length of the salt used in the hash in bytes. Increasing this value makes it more difficult for attackers to generate pre-computed tables or rainbow tables that can be used to reverse engineer the hash.

The values used for each of these parameters should be chosen based on the desired trade-off between security and performance. Higher values generally result in more secure hashes, but at the cost of longer computation times and increased memory usage.